Update Problem 3.cpp

+++ Problem 3.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main() {
 	
- double C,F;
+ float C,F;
  
  cout << "Enter temperature in Fahrenheit :" << endl;
  cin >> F;
